(trivial) fix some warnings
Acked-by: Stefan Reinauer <stepan@coresystems.de>



git-svn-id: svn://svn.coreboot.org/coreboot/trunk@4076 2b7e53f0-3cfb-0310-b3e9-8179ed1497e1

+++ src/arch/i386/boot/coreboot_table.c
@@ -204,7 +204,7 @@ struct lb_forward *lb_forward(struct lb_header *header, struct lb_header *next_h
 	forward = (struct lb_forward *)rec;
 	forward->tag = LB_TAG_FORWARD;
 	forward->size = sizeof(*forward);
-	forward->forward = (uint64_t) next_header;
+	forward->forward = (uint64_t)(unsigned long)next_header;
 	return forward;
 }
